Remove unnecessary routes, refactor files into packages
