+++ src/main/java/com/example/botfightwebserver/gameMatch/application/GameMatchRescheduler.java
@@ -1,5 +1,6 @@
-package com.example.botfightwebserver.gameMatch;
+package com.example.botfightwebserver.gameMatch.application;
 
+import com.example.botfightwebserver.gameMatch.domain.GameMatchJob;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Scheduled;
