+++ src/main/java/com/example/botfightwebserver/gameMatch/domain/MATCH_REASON.java
@@ -1,4 +1,4 @@
-package com.example.botfightwebserver.gameMatch;
+package com.example.botfightwebserver.gameMatch.domain;
 
 public enum MATCH_REASON {
     VALIDATION,
