+++ src/main/java/com/example/botfightwebserver/gameMatch/domain/MATCH_STATUS.java
@@ -1,4 +1,4 @@
-package com.example.botfightwebserver.gameMatch;
+package com.example.botfightwebserver.gameMatch.domain;
 
 public enum MATCH_STATUS {
     WAITING,
