+++ src/main/java/com/example/botfightwebserver/gameMatch/domain/MatchSubmissionRequest.java
@@ -1,4 +1,4 @@
-package com.example.botfightwebserver.gameMatch;
+package com.example.botfightwebserver.gameMatch.domain;
 
 import lombok.Data;
