+++ src/main/java/com/example/botfightwebserver/glicko/GlickoCalculator.java
@@ -1,6 +1,6 @@
 package com.example.botfightwebserver.glicko;
 
-import com.example.botfightwebserver.gameMatch.MATCH_STATUS;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_STATUS;
 import com.example.botfightwebserver.team.domain.Team;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
