+++ src/main/java/com/example/botfightwebserver/scrimmageMatch/ScrimmageMatch.java
@@ -1,6 +1,6 @@
 package com.example.botfightwebserver.scrimmageMatch;
 
-import com.example.botfightwebserver.gameMatch.GameMatch;
+import com.example.botfightwebserver.gameMatch.domain.GameMatch;
 import com.example.botfightwebserver.team.domain.Team;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
