+++ src/main/java/com/example/botfightwebserver/scrimmageMatch/ScrimmageMatchController.java
@@ -1,9 +1,9 @@
 package com.example.botfightwebserver.scrimmageMatch;
 
 import com.example.botfightwebserver.auth.domain.User;
-import com.example.botfightwebserver.gameMatch.GameMatch;
-import com.example.botfightwebserver.gameMatch.GameMatchService;
-import com.example.botfightwebserver.gameMatch.MATCH_REASON;
+import com.example.botfightwebserver.gameMatch.domain.GameMatch;
+import com.example.botfightwebserver.gameMatch.application.GameMatchService;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_REASON;
 import com.example.botfightwebserver.player.application.PlayerService;
 import com.example.botfightwebserver.submission.Submission;
 import com.example.botfightwebserver.team.domain.Team;
