+++ src/main/java/com/example/botfightwebserver/scrimmageMatch/ScrimmageMatchRepository.java
@@ -1,6 +1,6 @@
 package com.example.botfightwebserver.scrimmageMatch;
 
-import com.example.botfightwebserver.gameMatch.MATCH_STATUS;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_STATUS;
 import com.example.botfightwebserver.team.domain.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
