+++ src/main/java/com/example/botfightwebserver/searchEngine/SearchEngineController.java
@@ -1,7 +1,7 @@
 package com.example.botfightwebserver.searchEngine;
 
-import com.example.botfightwebserver.gameMatch.GameMatchDTO;
-import com.example.botfightwebserver.gameMatch.MATCH_REASON;
+import com.example.botfightwebserver.gameMatch.domain.GameMatchDTO;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_REASON;
 import com.example.botfightwebserver.team.domain.PublicTeamDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -34,19 +34,5 @@ public class SearchEngineController {
         return ResponseEntity.ok(teamSearchResult);
     }
 
-    @GetMapping("/match")
-    public ResponseEntity<Page<GameMatchDTO>> searchGame(
-        @RequestParam(required = false) String teamSearchParam,
-        @RequestParam(required = false) Long teamId,
-        @RequestParam(required = false) MATCH_REASON reason,
-        @RequestParam(required = false) String map,
-        @RequestParam(defaultValue = "0") int page,
-        @RequestParam(defaultValue = "10") int size) {
-        System.out.println("searchGame");
-        Pageable pageable = PageRequest.of(page, size);
-        Page<GameMatchDTO> gameSearchResult =
-            searchEngineService.searchGame(Optional.ofNullable(teamSearchParam), Optional.ofNullable(teamId),
-                Optional.ofNullable(reason), Optional.ofNullable(map), pageable);
-        return ResponseEntity.ok(gameSearchResult);
-    }
+
 }
