+++ src/main/java/com/example/botfightwebserver/team/domain/EditTeamDto.java
@@ -1,0 +1,9 @@
+package com.example.botfightwebserver.team.domain;
+
+import lombok.Value;
+
+@Value
+public class EditTeamDto {
+    String name;
+    String quote;
+}
