+++ src/main/java/com/example/botfightwebserver/team/domain/Team.java
@@ -1,6 +1,6 @@
 package com.example.botfightwebserver.team.domain;
 
-import com.example.botfightwebserver.gameMatch.GameMatch;
+import com.example.botfightwebserver.gameMatch.domain.GameMatch;
 import com.example.botfightwebserver.submission.Submission;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.annotations.VisibleForTesting;
