+++ src/main/java/com/example/botfightwebserver/tournament/TournamentGameMatch.java
@@ -1,8 +1,6 @@
 package com.example.botfightwebserver.tournament;
 
-import com.example.botfightwebserver.gameMatch.GameMatch;
-import jakarta.persistence.DiscriminatorColumn;
-import jakarta.persistence.DiscriminatorValue;
+import com.example.botfightwebserver.gameMatch.domain.GameMatch;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -11,7 +9,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.MapsId;
 import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
