+++ src/test/java/com/example/botfightwebserver/gameMatch/GameMatchTest.java
@@ -1,5 +1,8 @@
 package com.example.botfightwebserver.gameMatch;
 
+import com.example.botfightwebserver.gameMatch.domain.GameMatch;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_REASON;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_STATUS;
 import com.example.botfightwebserver.team.domain.Team;
 import com.example.botfightwebserver.submission.Submission;
 import org.junit.jupiter.api.BeforeEach;
