+++ src/test/java/com/example/botfightwebserver/gameMatch/MatchSubmissionRequestTest.java
@@ -1,5 +1,7 @@
 package com.example.botfightwebserver.gameMatch;
 
+import com.example.botfightwebserver.gameMatch.domain.MATCH_REASON;
+import com.example.botfightwebserver.gameMatch.domain.MatchSubmissionRequest;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
