+++ src/test/java/com/example/botfightwebserver/glicko/GlickoCalculatorTest.java
@@ -1,6 +1,6 @@
 package com.example.botfightwebserver.glicko;
 
-import com.example.botfightwebserver.gameMatch.MATCH_STATUS;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_STATUS;
 import com.example.botfightwebserver.team.domain.Team;
 import org.junit.jupiter.api.Test;
