+++ src/test/java/com/example/botfightwebserver/rabbitMQ/RabbitMQListenerTest.java
@@ -2,7 +2,7 @@ package com.example.botfightwebserver.rabbitMQ;
 
 import com.example.botfightwebserver.gameMatchResult.GameMatchResult;
 import com.example.botfightwebserver.gameMatchResult.GameMatchResultHandler;
-import com.example.botfightwebserver.gameMatch.MATCH_STATUS;
+import com.example.botfightwebserver.gameMatch.domain.MATCH_STATUS;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
