+++ src/test/java/com/example/botfightwebserver/submission/SubmissionControllerTest.java
@@ -1,9 +1,9 @@
 //package com.example.botfightwebserver.submission;
 //
-//import com.example.botfightwebserver.gameMatch.GameMatch;
-//import com.example.botfightwebserver.gameMatch.GameMatchJob;
-//import com.example.botfightwebserver.gameMatch.GameMatchService;
-//import com.example.botfightwebserver.gameMatch.MATCH_REASON;
+//import com.example.botfightwebserver.gameMatch.domain.GameMatch;
+//import com.example.botfightwebserver.gameMatch.domain.GameMatchJob;
+//import com.example.botfightwebserver.gameMatch.application.GameMatchService;
+//import com.example.botfightwebserver.gameMatch.domain.MATCH_REASON;
 //import com.example.botfightwebserver.rabbitMQ.RabbitMQService;
 //import com.example.botfightwebserver.team.domain.Team;
 //import org.junit.jupiter.api.Test;
